PLUGIN-1524 redesign the plugin properties to split up path field
Split up the path property into individual properties for server
type, host, port, username, and password. This is a simpler
experience than forcing the user to enter all the information
in a correctly formatted URI, and also allows us to handle
special characters better.

Since this is a backwards incompatible change, the plugin name
is changed so that it is effectively a new plugin. The old SFTP
source will not be able to be upgraded to the new one.