PLUGIN-1525 fix colon in password for sftp
fix the logic to extract the password to be a little better and
work properly if the password contains a colon. This does not
fix issues if the username contains a colon, and only fixes
issues for SFTP and not FTP. This is because the underlying
Hadoop FTPFileSystem also does not handle colons in username or
password properly, while the SFTPFileSystem does not have that
same bug.